requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-common-sql==1.8.0
apache-airflow-providers-sqlite==3.5.0
airflow-provider-duckdb==0.2.0
astro-sdk-python[duckdb, google]==1.7.0
apache-airflow-providers-apache-spark==4.2.0
apache-airflow-providers-docker==3.8.0
apache-airflow-providers-http==4.6.0

# webscraping
beautifulsoup4==4.12.2
requests==2.31.0

# testing
pytest==7.4.2
pytest-cov==4.1.0

# linting & formating
ruff==0.0.292

# logging
loguru==0.7.2

# env variables
python-dotenv==0.21.0

# data wrangling (etl)
polars==0.19.12

# URL manipulation
furl==2.1.3

# Datetime
pendulum==2.1.2

# EDA Profiling
ydata-profiling==4.6.1
